Created a custom hook to fetch the response from an api
